Refactor models properties

+++ src/main/java/com/paymentsystem/factory/AuthorizeTransactionFactory.java
@@ -3,10 +3,12 @@ package com.paymentsystem.factory;
 import com.paymentsystem.model.AuthorizeTransaction;
 import com.paymentsystem.model.Transactable;
 
+import java.math.BigDecimal;
+
 public class AuthorizeTransactionFactory extends TransactableFactory {
 
     @Override
-    public Transactable createTransactable(Transactable referredTransaction, String customerEmail, String customerPhone, double amount) {
+    public Transactable createTransactable(Transactable referredTransaction, String customerEmail, String customerPhone, BigDecimal amount) {
         return new AuthorizeTransaction(amount, customerEmail, customerPhone);
     }
 }
