+++ src/main/java/com/paymentsystem/factory/TransactableFactory.java
@@ -2,8 +2,10 @@ package com.paymentsystem.factory;
 
 import com.paymentsystem.model.Transactable;
 
+import java.math.BigDecimal;
+
 public abstract class TransactableFactory {
 
-    public abstract Transactable createTransactable(Transactable referredTransaction, String customerEmail, String customerPhone, double amount);
+    public abstract Transactable createTransactable(Transactable referredTransaction, String customerEmail, String customerPhone, BigDecimal amount);
 
 }
