+++ src/main/java/com/paymentsystem/model/AuthorizeTransaction.java
@@ -2,10 +2,12 @@ package com.paymentsystem.model;
 
 import jakarta.persistence.Entity;
 
+import java.math.BigDecimal;
+
 @Entity
 public class AuthorizeTransaction extends Transaction {
 
-    public AuthorizeTransaction(double amount, String customerEmail, String customerPhone) {
+    public AuthorizeTransaction(BigDecimal amount, String customerEmail, String customerPhone) {
         super(null, amount, customerEmail, customerPhone, "approved");
     }
