+++ src/main/java/com/paymentsystem/model/ChargeTransaction.java
@@ -2,10 +2,12 @@ package com.paymentsystem.model;
 
 import jakarta.persistence.Entity;
 
+import java.math.BigDecimal;
+
 @Entity
 public class ChargeTransaction extends Transaction {
 
-    public ChargeTransaction(Transaction referredTransaction, double amount, String customerEmail, String customerPhone) {
+    public ChargeTransaction(Transaction referredTransaction, BigDecimal amount, String customerEmail, String customerPhone) {
         super(referredTransaction, amount, customerEmail, customerPhone, "error");
         if ("approved".equals(referredTransaction.getStatus())) {
             setStatus("approved");
